refac: Creating ExpenseOverviewSpendBar, making easy to change expense bar in the future
