+++ lib/expenses/screens/ExpenseList.dart
@@ -1,7 +1,7 @@
 import 'package:finance_controlinator_mobile/expenses/domain/overviews/ExpenseBrief.dart';
 import 'package:finance_controlinator_mobile/expenses/domain/overviews/ExpenseOverview.dart';
-import 'package:finance_controlinator_mobile/expenses/domain/overviews/ExpensePartition.dart';
 import 'package:finance_controlinator_mobile/expenses/webclients/ExpenseWebClient.dart';
+import 'package:finance_controlinator_mobile/expenses/screens/overview/ExpenseOverviewSpendBar.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
@@ -43,7 +43,7 @@ class ExpenseListHeader extends StatelessWidget {
               ),
               Padding(
                 padding: EdgeInsets.only(top: 16),
-                child: ExpenseListHeaderBar(snapshot.data!.partitions),
+                child: ExpenseOverviewSpendBar(snapshot.data!.partitions),
               ),
             ],
           );
@@ -53,94 +53,6 @@ class ExpenseListHeader extends StatelessWidget {
   }
 }
 
-class ExpenseListHeaderBar extends StatelessWidget {
-  List<ExpensePartition> _partitions;
-
-  ExpenseListHeaderBar(this._partitions);
-
-  @override
-  Widget build(BuildContext context) {
-    _partitions.sort((x,y) => x.type.compareTo(y.type));
-    var partitionsWithSpend =
-    _partitions.where((element) => element.percent > 0).toList();
-
-    List<Color> colors = [
-      Colors.redAccent,
-      Colors.greenAccent,
-      Colors.blueAccent,
-      Colors.orangeAccent,
-      Colors.blueGrey,
-      Colors.purpleAccent,
-    ];
-    var descriptions = _partitions
-        .asMap()
-        .map((key, value) => MapEntry(
-            key,
-            Row(
-              children: [
-                Padding(
-                  padding: EdgeInsets.only(left: 4, right: 4),
-                  child: Text(value.type + ":"),
-                ),
-                Container(
-                    height: 20,
-                    width: 20,
-                    decoration: BoxDecoration(
-                        color: colors[key],
-                        borderRadius: BorderRadius.all(Radius.circular(100))))
-              ],
-            )))
-        .values
-        .toList();
-
-    return Column(
-      children: [
-        Row(
-            children: partitionsWithSpend
-                .asMap()
-                .map((key, value) {
-                  var border = key == 0
-                      ? BorderRadius.only(
-                          topLeft: Radius.circular(6),
-                          bottomLeft: Radius.circular(6))
-                      : key == partitionsWithSpend.length - 1
-                          ? BorderRadius.only(
-                              topRight: Radius.circular(6),
-                              bottomRight: Radius.circular(6))
-                          : BorderRadius.all(Radius.zero);
-
-                  return MapEntry(
-                      key,
-                      Expanded(
-                          flex: value.percent.toInt(),
-                          child: Container(
-                            height: 20,
-                            decoration: BoxDecoration(
-                                color: colors[key], borderRadius: border),
-                          )));
-                })
-                .values
-                .toList()),
-        Padding(
-            padding: EdgeInsets.only(
-              top: 8,
-              bottom: 8,
-            ),
-            child: Row(
-                mainAxisAlignment: MainAxisAlignment.spaceAround,
-                children: descriptions.take(3).toList())),
-        Padding(
-            padding: EdgeInsets.only(
-              top: 8,
-              bottom: 8,
-            ),
-            child: Row(
-                mainAxisAlignment: MainAxisAlignment.spaceAround,
-                children: descriptions.skip(3).take(3).toList()))
-      ],
-    );
-  }
-}
 
 class ExpenseListHeaderCards extends StatelessWidget {
   List<ExpenseBrief> expenseBriefs;
